requirements.txt
fastapi
uvicorn
requests
openai>=1.0.0,<2.0.0
python-dotenv
apscheduler
cachetools
pytest
